nsfw-checker-pro/requirements.txt
customtkinter>=5.2.0
onnxruntime-gpu>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
psutil>=5.9.0
GPUtil>=1.4.0
transformers>=4.45.0
torch>=2.0.0
